RF: Move generic json read/write functions into data.utils.
Beginning to work towards s3bids deprecation.
