+++ AFQ/data/utils.py
@@ -31,3 +31,41 @@ BUNDLE_MAT_2_PYTHON = \
      'RightILF': 'ILF_R', 'LeftILF': 'ILF_L',
      'Right SLF': 'SLF_R', 'Left SLF': 'SLF_L',
      'RightSLF': 'SLF_R', 'LeftSLF': 'SLF_L'}
+
+
+def write_json(fname, data):
+    """
+    Write data to JSON file.
+
+    Parameters
+    ----------
+    fname : str
+        Full path to the file to write.
+
+    data : dict
+        A dict containing the data to write.
+
+    Returns
+    -------
+    None
+    """
+    with open(fname, 'w') as ff:
+        json.dump(data, ff, default=lambda obj: "Not Serializable")
+
+
+def read_json(fname):
+    """
+    Read data from a JSON file.
+
+    Parameters
+    ----------
+    fname : str
+        Full path to the data-containing file
+
+    Returns
+    -------
+    dict
+    """
+    with open(fname, 'r') as ff:
+        out = json.load(ff)
+    return out
