+++ AFQ/tasks/segmentation.py
@@ -14,7 +14,7 @@ import AFQ.segmentation as seg
 from AFQ.utils.path import drop_extension
 import AFQ.utils.streamlines as aus
 from AFQ.tasks.utils import get_default_args
-from AFQ.data.s3bids import write_json
+from AFQ.data.utils import write_json
 import AFQ.api.bundle_dict as abd
 import AFQ.utils.streamlines as aus
 import AFQ.utils.volume as auv
